requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.4.0
scipy>=1.11.0
openpyxl>=3.1.0
groq>=0.5.0
google-generativeai>=0.5.0
reportlab>=4.1.0
python-docx>=1.1.0
tabulate>=0.9.0
